[dev-3099] moving AwardHistory to shared dir
